backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "mjc0042",email = "mjc0042@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-ninja (>=1.4.3,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "django-ninja-jwt (>=5.3.7,<6.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]

package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
